Add support for `export`
The `export` keyword causes Pony to mark a class and its functions as
reachable, even if they are not used within Pony code. It also
generates wrapper functions around these functions so that they can be
called without using the mangled names.

This allows a C function to call methods on Pony objects even if those
methods were not used in Pony code (otherwise they would be
unreachable and would not be included in the final executable).