Add serialization and deserialization of pointer fields.
Prior to this commit, attempting to serialize an object with a pointer
field would cause a run-time error. But now with custom serialization
we can support serializing and deserializing objects with pointer
fields.

When a pointer field is serialized, a value with all bits set high is
stored for the field, and when this field is deserialized it is set to
NULL. It is the responsiblity of the custom serializer to assign an
appropriate value to this field.